refactor and modularize code finally
